Display age for recent / upcoming birthdays
This commit introduces an age label in the description of generated
events. As this requires events to be generated for each year
individually, only recent birthdays and a few upcoming ones will receive
the age label: users can configure the interval in the settings.

Sadly, the calendar API draft cannot handle recurrence exceptions, even
when using ical directly. Technically, birthdays with age labels are
thus not recurrence exceptions, but separate events. The main recurring
event is amended with suitable EXDATE rules to prevent overlaps.

This commit also fixes a few stylistic issues and adds DTSTAMP
generation (which is mandated by the ical standard).