Make age generation interval more explicit
The age calculation range setting did confuse multiple users, as it
refers to the whole duration ages are calculated for (as opposed to the
"distance" from the current year, e.g. half of the duration). This
commit tries to avoid confusion by two changes:
1. The age calculation range is now relative to the current year, so a
   range of 1 refers to the current year only, and each increase by one
   adds one year in the past and the future.
2. The settings UI now displays the current range alongside the setting,
   so users can more directly see the effect of the setting.
As part of these changes, this commit also fixes issue #10.