Adapt to address book API change in TB 102.0b7
The fact that the vCard of a contact was exposed as '_vCard' a bug and not
intended behavior (correct is 'vCard'). In addition, Thudnerbird 102
will ship with support for legacy properties, which was missing in
earlier betas. Both aspects were addressed in Thunderbird 102.0b7.

This commit updates the add-on to reflect these changes; note that the
previous code continues to work in Thunderbird 102.0b7, falling back to
Birth* properties, which is expected to at least work until TB 114.