Implement age cutoff feature
This commit adds an option to set a cutoff birth year. If selected, the
add-on will ignore birth years at or before the cutoff year. This
permits to work around issues with other clients not properly handling
missing birth years, but switching to clients that do not produce
mangled data is still advisable.

Fixes #17