Bump compatibility for TB114
This commit bumps the add-ons manifest to support Thunderbird 114.0b1
as I tested that version in preparation for the next major release
(Thunderbird "Supernova" 115) without any major issues surfacing.

At the moment, I hope to be able to bump to 115.* without any additional
changes once the first beta versions of Thunderbird 115 are released.