Add unsupported build type
To more easily permit the add-on to be used on versions that are not
officially supported, this commit permits to build "unsupported" xpis
that do not include a max version guard.

Users that know what they are doing can use these builds to use the
add-on as far as technically possible, at their own risk.